+++ Blazor.Extensions.RCanvas/Blazor.Extensions.RCanvas/Canvas/DOMPoint.cs
@@ -8,13 +8,13 @@ namespace WebComponents.Components.Controls.Canvas;
 
 public class DOMPoint : DOMPointReadOnly {
     
-    public int? w { get; set; }
+    public double? w { get; set; }
     
-    public int? x { get; set; }
+    public double? x { get; set; }
     
-    public int? y { get; set; }
+    public double? y { get; set; }
     
-    public int? z { get; set; }
+    public double? z { get; set; }
 }
 
 public class DOMPointReadOnly {
@@ -30,13 +30,13 @@ public class DOMPointReadOnly {
         get; set;
     }
     
-    public int? w { get; set; }
+    public double? w { get; set; }
     
-    public int?  x{ get; set; }
+    public double?  x{ get; set; }
     
-    public int?  y{ get; set; }
+    public double?  y{ get; set; }
     
-    public int?  z{ get; set; }
+    public double?  z{ get; set; }
     
     public async Task<DOMPoint> MatrixTransformAsync(DOMMatrixInit matrix) {
         string functionName  = GetJsFunctionName();
